Turn off tagbar, tabline, and install fzf
The tagbar is nice for glancing at code structure, but it takes up a lot
of space.  I'm removing it.

The tabline looks nice and works well opening small numbers of buffers,
but ends up being a distraction when you have a large number of buffers
and you begin to worry about their order.

fzf is used to help find files to open and to help navigate buffers.
I've removed my hotkeys to navigate to the previous or next buffer to
break myself out of the habit of thinking about buffers being in a
certain order.