fish: simplify env vars, tab indent, & mark folds
Loading environment variables only worked when ~/.profile is kept
simple.  If statements in ~/.profile would not work and checking if a
path exists before adding it to PATH would make config.fish more complex
than it already was to support it.

Loading ~/.profile was meant to satisfy Emacs, and since I don't use it
anymore, I'm switching back to simply using pure Fish for environment
variables.