Prepare for bare-repository dotfiles
Preparing to attempt to use a bare git repository for dotfiles as
described here: https://www.atlassian.com/git/tutorials/dotfiles

Rearranged all files to be in the same directory path as they would be
if they were in the $HOME directory without any links.

BobTheFish prompt has been removed so we do not rely on submodules for
this.  Instead, the `fisher` package manager will be used to install it.

Removed the term/ directory and the script to install terminfo.

Turned freebsd-pkgs into a script that can be run from .local/bin/

Removed aerc configuration and some unused configurations in fish.

No longer run neovim-update.sh when running dotbot