fish: Switch to fundle & fix prompts on non-macOS
Fundle - https://github.com/danhper/fundle - is now used for package
management.  This replaces fisher.

Package for ssh agent is now installed.  This seems more reliable that
the alternative I was running.  The same agent is used when starting a
new shell in tmux.

pwrem is not run on non-macOS environments.

pwrem is updated to warn that non-macOS environments are not supported.