git: configure web view bind to localhost:4321
This configures the `git instaweb` to default to binding to the local
network interface on port 4321 using Python.

Sometimes it's darn useful to view your repository locally in a browser,
especially if you cannot connect to the remote web view and you want to
view the code you've deleted in your current branch.

Alternative is to use `git show <BRANCH>:path/to/file` which I pipe
through `bat -l go -p` to view Go code with syntax highlighting but
without the page numbers so I can copy and paste from the terminal.