pwrem: add -q option to check if passwd is stored
Running `pwrem -q` or `pwrem --query` will silently check if there is a
password stored.  It currently returns 0 if running on MacOS and if
there is a password stored to practice.

This enables me to add this to my shell startup:

	if pwrem -q
		pwrem
	end

This means only machines that support pwrem will prompt me to practice
my password if I currently have a password stored for practice.