Use go to symlink dotfiles; remove dotbot
Remove dotbot to no longer depend on Python or git submodules.

Unfortunately the `ln` command is not the same across systems.
Different flags are supported and the same flag can mean something
different on different operating systems.

Go is currently on all the systems I run so I wrote install.go which can
be run with:

    go run install.go

This will symlink all my dotfiles configured from a map at the top of
the file.