towebp: add resize and dry-run options; use batch
Added the --resize option to specify the geometry that gets passed to
`gm convert` with the default being 1024x> which means to resize
anything larger than 1024 wide to be 1024 pixels wide but keep the
aspect ratio.

Added the --dry-run option to print the commands being sent to `gm
batch` without running.

Switched to using `gm batch` to process multiple images.