fish: add the pj command to jump between projects
pj allows you to easily jump between your favourite directories in a
predictable manner. You tell pj where to look for your projects, and it
will allow you to jump to them easily with tab completion. It even
provides a convenient ability to open an editor in that directory from
anywhere.

Source:

	https://github.com/oh-my-fish/plugin-pj

Configuration:

Set the PROJECT_PATHS variable to a list of directories where projects
are stored.

	$ set -gx PROJECT_PATHS ~/workspace ~/src

Usage:

To jump to a project:

	$ pj foo-project

To open your editor set by EDITOR:

	$ pj open foo-project