nvim: Use fzf.vim instead of Telescope or CtrlP
I prefer to use fzf instead of Telescope or CtrlP in NeoVIM.

Telescope is too new and depends on the latest and greatest version of
NeoVIM.  Not all package managers are equal so I can't always have the
latest version of NeoVIM everywhere.

CtrlP is compatible in more environments, but the fuzzy suggestions are
not as helpful from the way I'm used to using fzf.

fzf.vim has been dependable for years for me and the only thing that
bothered me about it was installing it with Plug which left a mess in my
home directory.

fzf.vim is now managed by vipk.fish and committed to my dotfiles.

fzf itself is installed with Homebrew on my Mac.  I may need to update
my init.vim if I install fzf using a package manager on another
environment.