nvim: fzf main plugin now managed by vipk
We depend on the fzf binary to be installed via a package manager.  This
installed the necessary VIM plugin into config/nvim/pack/mypkg/start/fzf
so we do not need to source it from an arbitrary location.

The vipk command now fetches just the VIM plugin part, not the binary.

This lets us remove changing the rtp setting in config/nvim/init.vim to
source the plugin where it's installed by a package manager.  On FreeBSD
I noticed that the VIM plugin does not exist.